app/cron/requirements.txt
fastapi
apscheduler
asyncio
uvicorn
python-dotenv
SQLAlchemy
psycopg2